Move forEachField to be an underscored stdlib function
